+++ crates/apps/looper/looper-vst/src/ui/common/mod.rs
@@ -1,0 +1,1 @@
+pub mod parameter_view;
