+++ crates/apps/looper/looper-vst/src/ui/common/parameter_view/parameter_view_model.rs
